pyproject.toml
[tool.poetry]
name = "barsic-bot"
version = "1.0.0"
description = ""
authors = ["Ivan Bazhenov <bazhenov.in@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.4.1"
emoji = "^2.10.1"
pydantic-settings = "^2.2.1"
uvicorn = "^0.29.0"
aiogram-dialog = "^2.1.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
isort = "^5.13.2"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
